Avoid mis-killing unrelated process.
Since commit ea33b945b1 (frameworks/base), ActivityManager executes
killProcessGroup on a separated background thread.

It is possible that when ActivityManager kills pid P and posts to
kill its group, before killing group executes, other process/thread
is started with the same P. Then killing group executes will kill P
again. Especially happens on zygote's daemon thread that results
system restart.

By checking if there exists a proc path for the given pid has
different uid from given uid, it should safe the most of cases.

Test: an app fork self then swipe it from recent app

+++ libprocessgroup/processgroup.cpp
@@ -256,6 +256,8 @@ static int killProcessGroupOnce(uid_t uid, int initialPid, int signal)
 {
     int processes = 0;
     struct ctx ctx;
+    struct stat stats;
+    char path[16] = {0};
     pid_t pid;
 
     ctx.initialized = false;
@@ -268,6 +270,15 @@ static int killProcessGroupOnce(uid_t uid, int initialPid, int signal)
             LOG(WARNING) << "Yikes, we've been told to kill pid 0!  How about we don't do that?";
             continue;
         }
+
+        snprintf(path, sizeof(path), "/proc/%d", pid);
+        if (stat(path, &stats) == 0 && stats.st_uid != uid) {
+            // Avoid killing the pid which has been reused by other process.
+            LOG(WARNING) << "Skip killing pid " << pid << " with uid " << uid
+                         << " that has different current uid " << stats.st_uid;
+            continue;
+        }
+
         LOG(VERBOSE) << "Killing pid " << pid << " in uid " << uid
                      << " as part of process group " << initialPid;
         if (kill(pid, signal) == -1) {
